requirements.txt
kafka-python>=2.0.2
six>=1.16.0
kafka-python-ng
pandas
requests
apache-airflow-providers-apache-kafka==1.6.0